# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TrustHood Escrow API
  version: 0.1.0
  description: |
    REST API for the TrustHood Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints return a standardised envelope:
    ```
    {
      "data": [...],
      "page": 1, "limit": 20, "total": 42,
      "totalPages": 3, "hasNextPage": true, "hasPreviousPage": false
    }
    ```

    ## Environments
    - **Testnet**: `https://api.test.trusthood-escrow.example.com`
    - **Mainnet**: `https://api.trusthood-escrow.example.com`
  contact:
    name: TrustHood Escrow
    url: https://trusthood-escrow.example.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  x-repository: https://github.com/trusthood-escrow/trusthood-escrow
  x-docs-url: https://docs.trusthood-escrow.example.com
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.test.trusthood-escrow.example.com
    description: Testnet
  - url: https://api.trusthood-escrow.example.com
    description: Mainnet
...
security:
  - BearerAuth: []
paths:
...
  /api/disputes:
    get:
      tags: [Disputes]
      summary: List disputes with pagination
      operationId: listDisputes
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/users/{address}/suspend:
    post:
      tags: [Admin]
      summary: Suspend a user
      operationId: suspendUser
      security:
        - AdminAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: Terms of service violation
      responses:
        '200':
          description: User suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/users/{address}/ban:
    post:
      tags: [Admin]
      summary: Permanently ban a user
      operationId: banUser
      security:
        - AdminAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: User banned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/disputes/{id}/resolve:
    post:
      tags: [Admin]
      summary: Resolve an open dispute
      operationId: resolveDispute
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientAmount, freelancerAmount]
              properties:
                clientAmount:
                  type: string
                  description: XLM amount to release to client
                  example: '250'
                freelancerAmount:
                  type: string
                  description: XLM amount to release to freelancer
                  example: '750'
                notes:
                  type: string
                  description: Internal resolution notes
      responses:
        '200':
          description: Dispute resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  txHash:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/settings:
    get:
      tags: [Admin]
      summary: Read current platform settings
      operationId: getSettings
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Platform settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Admin]
      summary: Update platform settings
      operationId: updateSettings
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platformFeePercent:
                  type: number
                  minimum: 0
                  maximum: 100
                  example: 2.5
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    Dispute:
      type: object
      properties:
        id:
          type: string
        escrowId:
          type: string
        raisedBy:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [Open, Resolved]
        resolution:
          type: object
          nullable: true
          properties:
            clientAmount:
              type: string
            freelancerAmount:
              type: string
            notes:
              type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    PaginatedDisputes:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Dispute'
...
    PlatformSettings:
      type: object
      properties:
        platformFeePercent:
          type: number
          example: 2.5
        minEscrowAmountXlm:
          type: string
        maxEscrowAmountXlm:
          type: string
        escrowLockPeriodDays:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...